Remove Textures heavy parameters
Resolves #175
Relates #170
