+++ Modules/MetalModule/Sources/MetalModule/Models/TextureFactory.swift
@@ -1,0 +1,273 @@
+//
+//  TextureFactory.swift
+//  MetalModule
+//
+//  Created by max on 04.05.2026.
+//
+
+import ModelIO
+import MetalKit
+
+struct TextureFactory {
+    let material: MDLMaterial?
+    let textureLoader: MTKTextureLoader
+
+    // MARK: Internal methods
+
+    func texturePropertyValue(with semantic: MDLMaterialSemantic) -> MTLTexture? {
+        guard let property = textureProperty(with: semantic) else {
+            return nil
+        }
+
+        return texture(from: property.textureSamplerValue?.texture,
+                       semantic: semantic)
+    }
+
+    func opacityTexture() -> MTLTexture? {
+        guard let property = textureProperty(with: .opacity) else {
+            return nil
+        }
+
+        let opacitySource = property.textureSamplerValue?.texture
+        let baseColorSource = textureProperty(with: .baseColor)?
+            .textureSamplerValue?
+            .texture
+        if let opacitySource,
+           let baseColorSource,
+           texturesShareSource(opacitySource, baseColorSource) {
+            return nil
+        }
+
+        return texture(from: opacitySource, semantic: .opacity)
+    }
+
+    func usesBaseColorAlpha(hasSeparateOpacityTexture: Bool) -> Bool {
+        if hasSeparateOpacityTexture {
+            return false
+        }
+
+        guard
+            let property = textureProperty(with: .baseColor),
+            let baseColorSource = property.textureSamplerValue?.texture
+        else {
+            return false
+        }
+
+        if let opacitySource = textureProperty(with: .opacity)?
+            .textureSamplerValue?
+            .texture,
+           texturesShareSource(opacitySource, baseColorSource) {
+            return textureHasVisibleAlpha(baseColorSource)
+        }
+
+        return textureHasVisibleAlpha(baseColorSource)
+    }
+
+    func baseColorFactor() -> SIMD3<Float> {
+        guard let property = scalarProperty(with: .baseColor) else {
+            return SIMD3<Float>(repeating: 1)
+        }
+
+        if textureProperty(with: .baseColor) != nil,
+           isModelIODefaultBaseColorFactor(property) {
+            return SIMD3<Float>(repeating: 1)
+        }
+
+        return colorFactor(from: property,
+                           default: SIMD3<Float>(repeating: 1))
+    }
+
+    func scalarFactor(for semantic: MDLMaterialSemantic,
+                      default defaultValue: Float) -> Float {
+        let semanticProperties = properties(with: semantic)
+        if let scalarProperty = semanticProperties.first(where: { $0.textureSamplerValue?.texture == nil }) {
+            return scalarProperty.floatValue
+        }
+
+        if textureProperty(with: semantic) != nil {
+            return 1
+        }
+
+        return defaultValue
+    }
+
+    // MARK: Private methods
+
+    private func properties(with semantic: MDLMaterialSemantic) -> [MDLMaterialProperty] {
+        material?.properties(with: semantic) ?? []
+    }
+
+    private func textureProperty(with semantic: MDLMaterialSemantic) -> MDLMaterialProperty? {
+        properties(with: semantic).first { $0.textureSamplerValue?.texture != nil }
+    }
+
+    private func texture(from mdlTexture: MDLTexture?,
+                         semantic: MDLMaterialSemantic) -> MTLTexture? {
+        guard let mdlTexture else {
+            return nil
+        }
+
+        let maxDimension = max(Int(mdlTexture.dimensions.x), Int(mdlTexture.dimensions.y))
+        let shouldGenerateMipmaps = maxDimension > 1
+
+        return try? textureLoader.newTexture(texture: mdlTexture,
+                                             options: textureOptions(for: semantic,
+                                                                     generateMipmaps: shouldGenerateMipmaps))
+    }
+
+    private func scalarProperty(with semantic: MDLMaterialSemantic) -> MDLMaterialProperty? {
+        properties(with: semantic).first { $0.textureSamplerValue?.texture == nil }
+    }
+
+    private func colorFactor(for semantic: MDLMaterialSemantic,
+                             default defaultValue: SIMD3<Float>) -> SIMD3<Float> {
+        guard let property = scalarProperty(with: semantic) else {
+            return defaultValue
+        }
+
+        return colorFactor(from: property, default: defaultValue)
+    }
+
+    private func colorFactor(from property: MDLMaterialProperty,
+                             default defaultValue: SIMD3<Float>) -> SIMD3<Float> {
+        switch property.type {
+        case .float:
+            return SIMD3<Float>(repeating: property.floatValue)
+        case .float3:
+            let value = property.float3Value
+            return SIMD3<Float>(value.x, value.y, value.z)
+        case .float4, .color:
+            let value = property.float4Value
+            return SIMD3<Float>(value.x, value.y, value.z)
+        default:
+            return defaultValue
+        }
+    }
+
+    private func textureHasVisibleAlpha(_ mdlTexture: MDLTexture) -> Bool {
+        guard mdlTexture.hasAlphaValues,
+              let pixels = normalizedRGBAProbe(for: mdlTexture, maxDimension: 128) else {
+            return false
+        }
+
+        var nonOpaqueCount = 0
+        let pixelCount = pixels.count / 4
+        var alphaIndex = 3
+        while alphaIndex < pixels.count {
+            if pixels[alphaIndex] < 255 {
+                nonOpaqueCount += 1
+            }
+            alphaIndex += 4
+        }
+
+        let alphaCoverage = Float(nonOpaqueCount) / Float(pixelCount)
+        return alphaCoverage >= 0.02
+    }
+
+    private func normalizedRGBAProbe(for mdlTexture: MDLTexture,
+                                     maxDimension: Int) -> [UInt8]? {
+        guard mdlTexture.channelCount >= 4,
+              let unmanagedImage = mdlTexture.imageFromTexture() else {
+            return nil
+        }
+
+        // `imageFromTexture` is not annotated/created as a retained Core
+        // Foundation result, so taking retained ownership over-releases it.
+        let image = unmanagedImage.takeUnretainedValue()
+        let width = image.width
+        let height = image.height
+        guard width > 0, height > 0 else {
+            return nil
+        }
+
+        let scale = min(1, Double(maxDimension) / Double(max(width, height)))
+        let probeWidth = max(1, Int((Double(width) * scale).rounded(.up)))
+        let probeHeight = max(1, Int((Double(height) * scale).rounded(.up)))
+        let bytesPerPixel = 4
+        let bytesPerRow = probeWidth * bytesPerPixel
+        var pixels = [UInt8](repeating: 0, count: probeHeight * bytesPerRow)
+        let bitmapInfo = CGImageAlphaInfo.premultipliedLast.rawValue |
+        CGBitmapInfo.byteOrder32Big.rawValue
+        guard let context = CGContext(data: &pixels,
+                                      width: probeWidth,
+                                      height: probeHeight,
+                                      bitsPerComponent: 8,
+                                      bytesPerRow: bytesPerRow,
+                                      space: CGColorSpaceCreateDeviceRGB(),
+                                      bitmapInfo: bitmapInfo) else {
+            return nil
+        }
+
+        context.draw(image, in: CGRect(x: 0,
+                                       y: 0,
+                                       width: probeWidth,
+                                       height: probeHeight))
+
+        return pixels
+    }
+
+    private func texturesHaveMatchingProbe(_ lhs: MDLTexture, _ rhs: MDLTexture) -> Bool {
+        guard lhs.dimensions.x == rhs.dimensions.x,
+              lhs.dimensions.y == rhs.dimensions.y,
+              lhs.channelCount == rhs.channelCount,
+              lhs.channelEncoding == rhs.channelEncoding,
+              let lhsProbe = normalizedRGBAProbe(for: lhs, maxDimension: 32),
+              let rhsProbe = normalizedRGBAProbe(for: rhs, maxDimension: 32) else {
+            return false
+        }
+
+        return lhsProbe == rhsProbe
+    }
+
+    private func texturesShareSource(_ lhs: MDLTexture, _ rhs: MDLTexture) -> Bool {
+        if lhs === rhs {
+            return true
+        }
+
+        let lhsIdentifier = textureURLIdentifier(lhs)
+        let rhsIdentifier = textureURLIdentifier(rhs)
+        if let lhsIdentifier, let rhsIdentifier {
+            guard lhsIdentifier == rhsIdentifier else {
+                return false
+            }
+            return true
+        }
+
+        return texturesHaveMatchingProbe(lhs, rhs)
+    }
+
+    private func textureURLIdentifier(_ texture: MDLTexture) -> String? {
+        if let urlTexture = texture as? MDLURLTexture {
+            let url = urlTexture.url
+            if url.isFileURL {
+                return "url:\(url.standardizedFileURL.path)"
+            }
+            return "url:\(url.absoluteString)"
+        }
+
+        return nil
+    }
+
+    private func isModelIODefaultBaseColorFactor(_ property: MDLMaterialProperty) -> Bool {
+        guard property.name == "baseColor",
+              property.type == .float3 else {
+            return false
+        }
+
+        let value = property.float3Value
+        let defaultValue: Float = 0.18
+        let epsilon: Float = 0.0001
+        return abs(value.x - defaultValue) < epsilon &&
+        abs(value.y - defaultValue) < epsilon &&
+        abs(value.z - defaultValue) < epsilon
+    }
+
+    private func textureOptions(for semantic: MDLMaterialSemantic,
+                                generateMipmaps: Bool) -> [MTKTextureLoader.Option: NSNumber] {
+        let usesSRGB = semantic == .baseColor || semantic == .emission
+        return [
+            MTKTextureLoader.Option.generateMipmaps: NSNumber(value: generateMipmaps),
+            MTKTextureLoader.Option.SRGB: NSNumber(value: usesSRGB)
+        ]
+    }
+}
